+++ src/main/java/com/ohgiraffers/hw22thteamproject/recipe/command/infrastructure/repository/JpaDishRepository.java
@@ -4,5 +4,5 @@ import com.ohgiraffers.hw22thteamproject.recipe.command.domain.aggregate.Dish;
 import com.ohgiraffers.hw22thteamproject.recipe.command.domain.repository.DishRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaDishRepository extends DishRepository, JpaRepository<Dish,Long> {
+public interface JpaDishRepository extends DishRepository, JpaRepository<Dish,Integer> {
 }
