+++ src/main/java/com/ohgiraffers/hw22thteamproject/recipe/command/infrastructure/repository/JpaRecipeRepository.java
@@ -5,4 +5,5 @@ import com.ohgiraffers.hw22thteamproject.recipe.command.domain.aggregate.Recipe;
 import com.ohgiraffers.hw22thteamproject.recipe.command.domain.repository.RecipeRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaRecipeRepository extends RecipeRepository, JpaRepository<Recipe,Long> {}
+public interface JpaRecipeRepository extends RecipeRepository, JpaRepository<Recipe,Integer> {
+}
